+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [3.3.5] - 2024-04
+
+### Added
+
+- Validation of transit gateway route table names to improve error message in case of duplicate names
+
+### Changed
+
+- Removed dependency on 'requests' library to mitigate CVE-2024-3651
+
 ## [3.3.4] - 2024-04
 
 ### Fixed